Fix audio decoder to support files over 2GB
- Casts the encoded buffer size to int64_t instead of int in
  GenericAudioDecoder::OpenImpl to avoid overflow on files exceeding 2GB.
- Adds a corresponding test for that.

+++ dali/operators/decoder/audio/generic_decoder.cc
@@ -194,7 +194,7 @@ class GenericAudioDecoder : public AudioDecoderBase {
 
 AudioMetadata GenericAudioDecoder::OpenImpl(span<const char> encoded) {
   assert(!encoded.empty());
-  mem_stream_ = {encoded.data(), static_cast<int>(encoded.size())};
+  mem_stream_ = {encoded.data(), static_cast<int64_t>(encoded.size())};
   SF_VIRTUAL_IO sf_virtual_io = {
           &GetFileLen,
           &Seek,
